requirements.txt
# Core dependencies for hypothesis elimination engine backend
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.23.0

# LangChain + LangGraph for agent orchestration
langgraph>=0.0.20
langchain>=0.1.0
langchain-core>=0.1.0
langchain-google-genai>=0.0.5

# Environment and utilities
python-dotenv
python-multipart
sse-starlette
